Combine value checkpointing
