--- packages/arb-avm-cpp/data_storage/include/data_storage/checkpoint/machinestatefetcher.hpp
@@ -1,31 +1,0 @@
-/*
- * Copyright 2019, Offchain Labs, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#ifndef machinestatefetcher_hpp
-#define machinestatefetcher_hpp
-
-#include <avm_values/value.hpp>
-
-class Transaction;
-
-template <typename T>
-struct DbResult;
-
-DbResult<value> getValue(const Transaction& transaction,
-                         uint256_t value_hash,
-                         TuplePool* pool);
-
-#endif /* machinestatefetcher_hpp */
