+++ packages/arb-avm-cpp/data_storage/src/checkpoint/checkpointutils.cpp
@@ -25,9 +25,6 @@
 #include <bigint_utils.hpp>
 
 constexpr int UINT64_SIZE = 8;
-constexpr int TUP_TUPLE_LENGTH = 33;
-constexpr int TUP_NUM_LENGTH = 33;
-constexpr int TUP_CODEPT_LENGTH = 41;
 
 std::unordered_map<int, int> blockreason_type_length = {{0, 1},
                                                         {1, 1},
@@ -45,42 +42,6 @@ void marshal_uint64_t(uint64_t val, std::vector<unsigned char>& buf) {
     buf.insert(buf.end(), tmpbuf.begin(), tmpbuf.end());
 }
 
-struct ValueSerializer {
-    std::vector<unsigned char> operator()(const Tuple& val) const {
-        std::vector<unsigned char> value_vector;
-        value_vector.push_back(TUPLE);
-        auto hash_key = hash_value(val);
-        marshal_uint256_t(hash_key, value_vector);
-
-        return value_vector;
-    }
-
-    std::vector<unsigned char> operator()(const uint256_t& val) const {
-        std::vector<unsigned char> value_vector;
-        value_vector.push_back(NUM);
-        marshal_uint256_t(val, value_vector);
-
-        return value_vector;
-    }
-
-    std::vector<unsigned char> operator()(const CodePointStub& val) const {
-        std::vector<unsigned char> value_vector;
-        auto type_code = static_cast<unsigned char>(CODEPT);
-        value_vector.push_back(type_code);
-        utils::serializeCodePointStub(val, value_vector);
-        return value_vector;
-    }
-
-    std::vector<unsigned char> operator()(const HashPreImage& val) const {
-        std::vector<unsigned char> value_vector;
-        auto type_code = static_cast<unsigned char>(HASH_PRE_IMAGE);
-        value_vector.push_back(type_code);
-        val.marshal(value_vector);
-
-        return value_vector;
-    }
-};
-
 namespace utils {
 
 void serializeCodePointStub(const CodePointStub& val,
@@ -97,53 +58,5 @@ uint64_t deserialize_uint64(const char*& bufptr) {
     return val;
 }
 
-std::vector<std::vector<unsigned char>> parseTuple(
-    const std::vector<unsigned char>& data) {
-    std::vector<std::vector<unsigned char>> return_vector;
-
-    auto iter = data.begin();
-    uint8_t count = *iter - TUPLE;
-    ++iter;
-
-    for (uint8_t i = 0; i < count; i++) {
-        auto value_type = static_cast<ValueTypes>(*iter);
-        std::vector<unsigned char> current;
-
-        switch (value_type) {
-            case NUM: {
-                auto next_it = iter + TUP_NUM_LENGTH;
-                current.insert(current.end(), iter, next_it);
-                iter = next_it;
-                break;
-            }
-            case CODEPT: {
-                auto next_it = iter + TUP_CODEPT_LENGTH;
-                current.insert(current.end(), iter, next_it);
-                iter = next_it;
-                break;
-            }
-            case HASH_PRE_IMAGE: {
-                throw std::runtime_error("HASH_ONLY item");
-            }
-            default: {
-                uint8_t tup_size = value_type - TUPLE;
-                if (tup_size > 8) {
-                    throw std::runtime_error(
-                        "tried to parse tuple with invalid typecode");
-                }
-                auto next_it = iter + TUP_TUPLE_LENGTH;
-                current.insert(current.end(), iter, next_it);
-                iter = next_it;
-                break;
-            }
-        }
-        return_vector.push_back(current);
-    }
-    return return_vector;
-}
-
-std::vector<unsigned char> serializeValue(const value& val) {
-    return nonstd::visit(ValueSerializer{}, val);
-}
 }  // namespace utils
 }  // namespace checkpoint
