+++ packages/arb-avm-cpp/tests/machine.cpp
@@ -20,7 +20,7 @@
 #include <avm/machine.hpp>
 #include <data_storage/checkpoint/checkpointstorage.hpp>
 #include <data_storage/checkpoint/machine.hpp>
-#include <data_storage/checkpoint/machinestatedeleter.hpp>
+#include <data_storage/checkpoint/value.hpp>
 #include <data_storage/storageresult.hpp>
 
 #include <catch2/catch.hpp>
